Move normalizer to pkg/trace/traceutil
