+++ pkg/trace/agent/normalizer_test.go
@@ -6,17 +6,16 @@
 package agent
 
 import (
-	"bytes"
 	"fmt"
 	"math"
 	"math/rand"
 	"strings"
 	"testing"
 	"time"
-	"unicode"
 
 	"github.com/DataDog/datadog-agent/pkg/trace/info"
 	"github.com/DataDog/datadog-agent/pkg/trace/pb"
+	"github.com/DataDog/datadog-agent/pkg/trace/traceutil"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -77,7 +76,7 @@ func TestNormalizeEmptyServiceNoLang(t *testing.T) {
 	s := newTestSpan()
 	s.Service = ""
 	assert.NoError(t, normalize(ts, s))
-	assert.Equal(t, DefaultServiceName, s.Service)
+	assert.Equal(t, traceutil.DefaultServiceName, s.Service)
 	assert.Equal(t, tsMalformed(&info.SpansMalformed{ServiceEmpty: 1}), ts)
 }
 
@@ -98,7 +97,7 @@ func TestNormalizeLongService(t *testing.T) {
 	s := newTestSpan()
 	s.Service = strings.Repeat("CAMEMBERT", 100)
 	assert.NoError(t, normalize(ts, s))
-	assert.Equal(t, s.Service, s.Service[:MaxServiceLen])
+	assert.Equal(t, s.Service, s.Service[:traceutil.MaxServiceLen])
 	assert.Equal(t, tsMalformed(&info.SpansMalformed{ServiceTruncate: 1}), ts)
 }
 
@@ -116,7 +115,7 @@ func TestNormalizeEmptyName(t *testing.T) {
 	s := newTestSpan()
 	s.Name = ""
 	assert.NoError(t, normalize(ts, s))
-	assert.Equal(t, s.Name, DefaultSpanName)
+	assert.Equal(t, s.Name, traceutil.DefaultSpanName)
 	assert.Equal(t, tsMalformed(&info.SpansMalformed{SpanNameEmpty: 1}), ts)
 }
 
@@ -125,7 +124,7 @@ func TestNormalizeLongName(t *testing.T) {
 	s := newTestSpan()
 	s.Name = strings.Repeat("CAMEMBERT", 100)
 	assert.NoError(t, normalize(ts, s))
-	assert.Equal(t, s.Name, s.Name[:MaxNameLen])
+	assert.Equal(t, s.Name, s.Name[:traceutil.MaxNameLen])
 	assert.Equal(t, tsMalformed(&info.SpansMalformed{SpanNameTruncate: 1}), ts)
 }
 
@@ -134,7 +133,7 @@ func TestNormalizeNameNoAlphanumeric(t *testing.T) {
 	s := newTestSpan()
 	s.Name = "/"
 	assert.NoError(t, normalize(ts, s))
-	assert.Equal(t, s.Name, DefaultSpanName)
+	assert.Equal(t, s.Name, traceutil.DefaultSpanName)
 	assert.Equal(t, tsMalformed(&info.SpansMalformed{SpanNameInvalid: 1}), ts)
 }
 
@@ -432,81 +431,4 @@ func BenchmarkNormalization(b *testing.B) {
 
 		normalize(ts, span)
 	}
-}
-
-func TestNormalizeTag(t *testing.T) {
-	for _, tt := range []struct{ in, out string }{
-		{in: "#test_starting_hash", out: "test_starting_hash"},
-		{in: "TestCAPSandSuch", out: "testcapsandsuch"},
-		{in: "Test Conversion Of Weird !@#$%^&**() Characters", out: "test_conversion_of_weird_characters"},
-		{in: "$#weird_starting", out: "weird_starting"},
-		{in: "allowed:c0l0ns", out: "allowed:c0l0ns"},
-		{in: "1love", out: "love"},
-		{in: "ünicöde", out: "ünicöde"},
-		{in: "ünicöde:metäl", out: "ünicöde:metäl"},
-		{in: "Data🐨dog🐶 繋がっ⛰てて", out: "data_dog_繋がっ_てて"},
-		{in: " spaces   ", out: "spaces"},
-		{in: " #hashtag!@#spaces #__<>#  ", out: "hashtag_spaces"},
-		{in: ":testing", out: ":testing"},
-		{in: "_foo", out: "foo"},
-		{in: ":::test", out: ":::test"},
-		{in: "contiguous_____underscores", out: "contiguous_underscores"},
-		{in: "foo_", out: "foo"},
-		{in: "\u017Fodd_\u017Fcase\u017F", out: "\u017Fodd_\u017Fcase\u017F"}, // edge-case
-		{in: "", out: ""},
-		{in: " ", out: ""},
-		{in: "ok", out: "ok"},
-		{in: "™Ö™Ö™™Ö™", out: "ö_ö_ö"},
-		{in: "AlsO:ök", out: "also:ök"},
-		{in: ":still_ok", out: ":still_ok"},
-		{in: "___trim", out: "trim"},
-		{in: "12.:trim@", out: ":trim"},
-		{in: "12.:trim@@", out: ":trim"},
-		{in: "fun:ky__tag/1", out: "fun:ky_tag/1"},
-		{in: "fun:ky@tag/2", out: "fun:ky_tag/2"},
-		{in: "fun:ky@@@tag/3", out: "fun:ky_tag/3"},
-		{in: "tag:1/2.3", out: "tag:1/2.3"},
-		{in: "---fun:k####y_ta@#g/1_@@#", out: "fun:k_y_ta_g/1"},
-		{in: "AlsO:œ#@ö))œk", out: "also:œ_ö_œk"},
-		{in: "test\x99\x8faaa", out: "test_aaa"},
-		{in: "test\x99\x8f", out: "test"},
-		{in: strings.Repeat("a", 888), out: strings.Repeat("a", 200)},
-		{
-			in: func() string {
-				b := bytes.NewBufferString("a")
-				for i := 0; i < 799; i++ {
-					_, err := b.WriteRune('🐶')
-					assert.NoError(t, err)
-				}
-				_, err := b.WriteRune('b')
-				assert.NoError(t, err)
-				return b.String()
-			}(),
-			out: "a", // 'b' should have been truncated
-		},
-		{"a" + string(unicode.ReplacementChar), "a"},
-		{"a" + string(unicode.ReplacementChar) + string(unicode.ReplacementChar), "a"},
-		{"a" + string(unicode.ReplacementChar) + string(unicode.ReplacementChar) + "b", "a_b"},
-	} {
-		t.Run("", func(t *testing.T) {
-			assert.Equal(t, tt.out, normalizeTag(tt.in), tt.in)
-		})
-	}
-}
-
-func benchNormalizeTag(tag string) func(b *testing.B) {
-	return func(b *testing.B) {
-		b.ReportAllocs()
-		for i := 0; i < b.N; i++ {
-			normalizeTag(tag)
-		}
-	}
-}
-
-func BenchmarkNormalizeTag(b *testing.B) {
-	b.Run("ok", benchNormalizeTag("good_tag"))
-	b.Run("trim", benchNormalizeTag("___trim_left"))
-	b.Run("trim-both", benchNormalizeTag("___trim_right@@#!"))
-	b.Run("plenty", benchNormalizeTag("fun:ky_ta@#g/1"))
-	b.Run("more", benchNormalizeTag("fun:k####y_ta@#g/1_@@#"))
 }
