+++ pkg/trace/agent/truncator_test.go
@@ -10,6 +10,7 @@ import (
 	"testing"
 
 	"github.com/DataDog/datadog-agent/pkg/trace/pb"
+	"github.com/DataDog/datadog-agent/pkg/trace/traceutil"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -62,7 +63,7 @@ func TestTruncateMetricsKeyTooLong(t *testing.T) {
 	s.Metrics[key] = 42
 	Truncate(s)
 	for k := range s.Metrics {
-		assert.True(t, len(k) < MaxMetricsKeyLen+4)
+		assert.True(t, len(k) < traceutil.MaxMetricsKeyLen+4)
 	}
 }
 
@@ -79,7 +80,7 @@ func TestTruncateMetaKeyTooLong(t *testing.T) {
 	s.Meta[key] = "foo"
 	Truncate(s)
 	for k := range s.Meta {
-		assert.True(t, len(k) < MaxMetaKeyLen+4)
+		assert.True(t, len(k) < traceutil.MaxMetaKeyLen+4)
 	}
 }
 
@@ -89,6 +90,6 @@ func TestTruncateMetaValueTooLong(t *testing.T) {
 	s.Meta["foo"] = val
 	Truncate(s)
 	for _, v := range s.Meta {
-		assert.True(t, len(v) < MaxMetaValLen+4)
+		assert.True(t, len(v) < traceutil.MaxMetaValLen+4)
 	}
 }
